Consolidate sharing features, reduce complexity, and liquidate tech debt
- Refactored high-complexity components: `MedicalAnalysisService`, `VitalSignAnalyzer`, `DeviceCapabilityService`.
- Consolidated `ble_sharing` and `health_sharing` features into a unified `health_sharing` module.
- Introduced `ProtocolHandler` to unify BLE/NFC/WiFi state handling across sharing Cubits.
- Liquidated 4 identified tech debt items (TODOs/XXXs) including repo integration and SSI versioning.
- Implemented 11+ new unit tests for medical analysis strategies and verified full test suite.
- Hardened dependencies: upgraded `flutter_gemma` and `flutter_secure_storage`.
- Prepared for DI hardening by adding `@lazySingleton` annotations to core services.
