+++ lib/features/health_sharing/application/protocol_handler.dart
@@ -1,32 +1,9 @@
-import 'package:equatable/equatable.dart';
-import '../../ble_sharing/domain/ble_sharing_service.dart' as ble;
-import '../../ble_sharing/infrastructure/nfc_sharing_service.dart' as ble_nfc;
-import '../../ble_sharing/infrastructure/wifi_direct_service.dart' as ble_wifi;
-import '../infrastructure/ble_sharing_service.dart' as health_ble;
+import "../infrastructure/ble_sharing_service.dart" as ble;
 import '../infrastructure/nfc_sharing_service.dart' as nfc;
 import '../infrastructure/wifi_direct_service.dart' as wifi;
 
-abstract class ProtocolState extends Equatable {
-  final String status;
-  final String? message;
-  final bool isError;
-  final int? bytesTransferred;
-  final Duration? transferTime;
-
-  const ProtocolState({
-    required this.status,
-    this.message,
-    this.isError = false,
-    this.bytesTransferred,
-    this.transferTime,
-  });
-
-  @override
-  List<Object?> get props => [status, message, isError, bytesTransferred, transferTime];
-}
-
 class ProtocolHandler {
-  static void handleBleState(ble.BleServiceState state, Function(ProtocolEvent) onEvent) {
+  static void handleHealthBleState(ble.BleServiceState state, Function(ProtocolEvent) onEvent) {
     if (state.status == 'scanning') {
       onEvent(ProtocolEvent.scanning);
     } else if (state.status == 'advertising') {
@@ -47,28 +24,7 @@ class ProtocolHandler {
     }
   }
 
-  static void handleHealthBleState(health_ble.BleSharingState state, Function(ProtocolEvent) onEvent) {
-    if (state.status == 'scanning') {
-      onEvent(ProtocolEvent.scanning);
-    } else if (state.status == 'advertising') {
-      onEvent(ProtocolEvent.advertising(state.deviceId ?? ''));
-    } else if (state.status == 'connecting') {
-      onEvent(ProtocolEvent.connecting(state.deviceId ?? ''));
-    } else if (state.status == 'connected') {
-      onEvent(ProtocolEvent.connected(state.deviceId ?? ''));
-    } else if (state.status == 'transferring') {
-      onEvent(ProtocolEvent.transferring(0.5, state.message ?? 'Transferring...'));
-    } else if (state.status == 'completed') {
-      onEvent(ProtocolEvent.completed(
-        state.bytesTransferred ?? 0,
-        state.transferTime ?? Duration.zero,
-      ));
-    } else if (state.isError) {
-      onEvent(ProtocolEvent.error(state.message ?? 'BLE Error'));
-    }
-  }
-
-  static void handleNfcState(nfc.NfcSharingState state, Function(ProtocolEvent) onEvent) {
+  static void handleNfcState(nfc.NfcHealthSharingState state, Function(ProtocolEvent) onEvent) {
     if (state.status == 'listening') {
       onEvent(ProtocolEvent.scanning);
     } else if (state.status == 'ndef_beam') {
@@ -85,45 +41,7 @@ class ProtocolHandler {
     }
   }
 
-  static void handleBleNfcState(ble_nfc.NfcSharingState state, Function(ProtocolEvent) onEvent) {
-    if (state.status == 'listening') {
-      onEvent(ProtocolEvent.scanning);
-    } else if (state.status == 'ndef_beam') {
-      onEvent(ProtocolEvent.transferring(0.5, state.message ?? 'Beaming...'));
-    } else if (state.status == 'received') {
-      onEvent(ProtocolEvent.received(state.receivedPackage));
-    } else if (state.status == 'completed') {
-      onEvent(ProtocolEvent.completed(
-        state.bytesTransferred ?? 0,
-        state.transferTime ?? Duration.zero,
-      ));
-    } else if (state.isError) {
-      onEvent(ProtocolEvent.error(state.message ?? 'NFC Error'));
-    }
-  }
-
-  static void handleWifiState(wifi.WifiSharingState state, Function(ProtocolEvent) onEvent) {
-    if (state.status == 'discovering') {
-      onEvent(ProtocolEvent.scanning);
-    } else if (state.status == 'hosting') {
-      onEvent(ProtocolEvent.advertising(state.address ?? ''));
-    } else if (state.status == 'connecting') {
-      onEvent(ProtocolEvent.connecting(state.address ?? ''));
-    } else if (state.status == 'transferring') {
-      onEvent(ProtocolEvent.transferring(0.5, state.message ?? 'Transferring...'));
-    } else if (state.status == 'received') {
-      onEvent(ProtocolEvent.received(state.receivedPackage));
-    } else if (state.status == 'completed') {
-      onEvent(ProtocolEvent.completed(
-        state.bytesTransferred ?? 0,
-        state.transferTime ?? Duration.zero,
-      ));
-    } else if (state.isError) {
-      onEvent(ProtocolEvent.error(state.message ?? 'WiFi Error'));
-    }
-  }
-
-  static void handleBleWifiState(ble_wifi.WifiServiceState state, Function(ProtocolEvent) onEvent) {
+  static void handleWifiState(wifi.WifiServiceState state, Function(ProtocolEvent) onEvent) {
     if (state.status == 'discovering') {
       onEvent(ProtocolEvent.scanning);
     } else if (state.status == 'hosting') {
