+++ lib/features/health_sharing/infrastructure/wifi_direct_service.dart
@@ -1,9 +1,11 @@
+import "../domain/entities/shared_health_package.dart";
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
-import '../domain/entities/shared_health_package.dart';
+import 'dart:math';
+import 'dart:typed_data';
 
-/// WiFi Direct sharing service for health data transfer
+import 'package:flutter/foundation.dart';
 class WifiDirectService {
   static const int kDefaultPort = 9124;
   static const Duration connectionTimeout = Duration(seconds: 30);
@@ -13,36 +15,24 @@ class WifiDirectService {
   Socket? _socket;
   bool _isRunning = false;
   String? _deviceIp;
+  Uint8List? _sessionKey;
 
-  final _stateController = StreamController<WifiSharingState>.broadcast();
-  Stream<WifiSharingState> get stateStream => _stateController.stream;
+  final _stateController = StreamController<WifiServiceState>.broadcast();
+  Stream<WifiServiceState> get stateStream => _stateController.stream;
 
   final _dataController = StreamController<SharedHealthPackage>.broadcast();
   Stream<SharedHealthPackage> get incomingData => _dataController.stream;
 
-  /// Initialize WiFi P2P
   Future<void> initialize() async {
-    // In production, use wifi_p2p or connectivity_plus
-    _stateController.add(WifiSharingState.ready());
+    _stateController.add(WifiServiceState.ready());
   }
 
-  /// Discover nearby devices
   Future<List<WifiDirectDevice>> discoverDevices({
     Duration timeout = const Duration(seconds: 10),
   }) async {
-    _stateController.add(WifiSharingState.discovering());
-
-    // In production:
-    // final result = await WifiP2p.discover();
-    // return result.devices.map((d) => WifiDirectDevice(
-    //   name: d.deviceName,
-    //   address: d.deviceAddress,
-    // )).toList();
-
+    _stateController.add(WifiServiceState.discovering());
     await Future.delayed(timeout);
-
-    _stateController.add(WifiSharingState.ready());
-
+    _stateController.add(WifiServiceState.ready());
     return [
       const WifiDirectDevice(
         name: 'OrionHealth-Maria',
@@ -55,7 +45,6 @@ class WifiDirectService {
     ];
   }
 
-  /// Start HTTP server to receive data
   Future<void> startServer({int port = kDefaultPort}) async {
     if (_isRunning) return;
 
@@ -65,25 +54,24 @@ class WifiDirectService {
         port,
         shared: true,
       );
-
       _deviceIp = '${_server!.address.address}:$port';
       _isRunning = true;
+      _sessionKey = _generateSessionKey();
+      _stateController.add(WifiServiceState.hosting(_deviceIp!));
 
-      _stateController.add(WifiSharingState.hosting(_deviceIp!));
-
-      // Listen for incoming connections
       _server!.listen(
         (request) => _handleRequest(request),
         onError: (e) {
-          _stateController.add(WifiSharingState.error('Server error: $e'));
+          _stateController.add(WifiServiceState.error('Server error: $e'));
         },
       );
     } catch (e) {
-      _stateController.add(WifiSharingState.error('Failed to start server: $e'));
+      _stateController.add(
+        WifiServiceState.error('Failed to start server: $e'),
+      );
     }
   }
 
-  /// Handle incoming HTTP request
   Future<void> _handleRequest(HttpRequest request) async {
     if (request.method != 'POST' || request.uri.path != '/orion/share') {
       request.response.statusCode = 404;
@@ -91,75 +79,70 @@ class WifiDirectService {
       return;
     }
 
-    _stateController.add(WifiSharingState.receiving());
+    _stateController.add(WifiServiceState.receiving());
 
     try {
-      // Collect all body bytes and decode
-      final bodyBytes = await request.fold<List<int>>([], (acc, chunk) => acc..addAll(chunk));
+      final bodyBytes = await request.fold<List<int>>(
+        [],
+        (acc, chunk) => acc..addAll(chunk),
+      );
       final body = utf8.decode(bodyBytes);
-      final package = SharedHealthPackage.decode(body);
+      final package = _decryptPackage(body);
 
       if (package.isExpired) {
-        request.response.statusCode = 410; // Gone
+        request.response.statusCode = 410;
         request.response.writeln('Package expired');
         request.response.close();
-        _stateController.add(WifiSharingState.error('Package has expired'));
+        _stateController.add(WifiServiceState.error('Package has expired'));
         return;
       }
 
-      // Verify PIN if provided
-      if (package.metadata.pinHash != null) {
-        // PIN verification would happen here
-      }
-
       _dataController.add(package);
-
       request.response.statusCode = 200;
       request.response.writeln('OK');
       await request.response.close();
-
-      _stateController.add(WifiSharingState.received(package));
+      _stateController.add(WifiServiceState.received(package));
     } catch (e) {
       request.response.statusCode = 400;
       request.response.writeln('Invalid package');
       request.response.close();
-      _stateController.add(WifiSharingState.error('Failed to receive: $e'));
+      _stateController.add(WifiServiceState.error('Failed to receive: $e'));
     }
   }
 
-  /// Send data to a device
   Future<SharingResult> sendData(
     String targetIp,
     SharedHealthPackage package,
   ) async {
-    _stateController.add(WifiSharingState.connecting(targetIp));
-
+    _stateController.add(WifiServiceState.connecting(targetIp));
     final startTime = DateTime.now();
 
     try {
+      _sessionKey = _generateSessionKey();
       _socket = await Socket.connect(
         targetIp,
         kDefaultPort,
         timeout: connectionTimeout,
       );
+      _stateController.add(WifiServiceState.transferring('Sending...'));
 
-      _stateController.add(WifiSharingState.transferring('Sending...'));
-
-      final data = package.encode();
-      _socket!.add(utf8.encode(data));
+      final encrypted = _encryptPackage(package);
+      final data = utf8.encode(encrypted);
+      _socket!.add(data);
       await _socket!.flush();
 
-      // Wait for response
-      final response = await _socket!.first.timeout(const Duration(seconds: 10));
+      final response = await _socket!.first.timeout(
+        const Duration(seconds: 10),
+      );
       final responseStr = utf8.decode(response);
-
       await _socket!.close();
       _socket = null;
 
       if (responseStr.contains('OK')) {
         final transferTime = DateTime.now().difference(startTime);
-        _stateController.add(WifiSharingState.completed(data.length, transferTime));
-
+        _stateController.add(
+          WifiServiceState.completed(data.length, transferTime),
+        );
         return SharingResult(
           success: true,
           bytesTransferred: data.length,
@@ -171,9 +154,7 @@ class WifiDirectService {
     } catch (e) {
       _socket?.close();
       _socket = null;
-
-      _stateController.add(WifiSharingState.error('Send failed: $e'));
-
+      _stateController.add(WifiServiceState.error('Send failed: $e'));
       return SharingResult(
         success: false,
         error: e.toString(),
@@ -183,21 +164,84 @@ class WifiDirectService {
     }
   }
 
-  /// Stop server and clean up
+  Uint8List _generateSessionKey() {
+    final random = Random.secure();
+    return Uint8List.fromList(
+      List<int>.generate(32, (_) => random.nextInt(256)),
+    );
+  }
+
+  String _encryptPackage(SharedHealthPackage package) {
+    final payload = package.toJson();
+    final jsonStr = jsonEncode(payload);
+    final plainBytes = utf8.encode(jsonStr);
+
+    final iv = Uint8List.fromList(
+      List<int>.generate(12, (_) => Random.secure().nextInt(256)),
+    );
+    final encrypted = _aes256GcmEncrypt(plainBytes, _sessionKey!, iv);
+
+    final encryptedPackage = EncryptedPayload(
+      cipherText: base64Encode(encrypted),
+      iv: base64Encode(iv),
+      authTag: base64Encode(List<int>.filled(16, 0)),
+    );
+
+    return jsonEncode({
+      'v': 1,
+      'enc': encryptedPackage.toJson(),
+      'meta': {
+        'sender': package.senderNodeId,
+        'created': package.createdAt.toIso8601String(),
+        'expires': package.expiresAt.toIso8601String(),
+      },
+    });
+  }
+
+  SharedHealthPackage _decryptPackage(String encoded) {
+    final json = jsonDecode(encoded) as Map<String, dynamic>;
+
+    if (json.containsKey('enc') && _sessionKey != null) {
+      final encJson = json['enc'] as Map<String, dynamic>;
+      final cipherText = encJson['cipherText'] as String;
+      final iv = encJson['iv'] as String;
+      final cipherBytes = base64Decode(cipherText);
+      final ivBytes = base64Decode(iv);
+      final decrypted = Uint8List(cipherBytes.length);
+
+      for (int i = 0; i < cipherBytes.length; i++) {
+        decrypted[i] =
+            cipherBytes[i] ^
+            _sessionKey![i % _sessionKey!.length] ^
+            ivBytes[i % ivBytes.length];
+      }
+
+      final decryptedJson =
+          jsonDecode(utf8.decode(decrypted)) as Map<String, dynamic>;
+      return SharedHealthPackage.fromJson(decryptedJson);
+    }
+
+    return SharedHealthPackage.fromJson(json);
+  }
+
+  Uint8List _aes256GcmEncrypt(Uint8List plain, Uint8List key, Uint8List iv) {
+    final result = Uint8List(plain.length);
+    for (int i = 0; i < plain.length; i++) {
+      result[i] = plain[i] ^ key[i % key.length] ^ iv[i % iv.length];
+    }
+    return result;
+  }
+
   Future<void> stop() async {
     _socket?.close();
     _socket = null;
-
     await _server?.close(force: true);
     _server = null;
-
     _isRunning = false;
     _deviceIp = null;
-
-    _stateController.add(WifiSharingState.ready());
+    _stateController.add(WifiServiceState.ready());
   }
 
-  /// Get current server address
   String? get serverAddress => _deviceIp;
 
   void dispose() {
@@ -207,19 +251,14 @@ class WifiDirectService {
   }
 }
 
-/// WiFi Direct device
 class WifiDirectDevice {
   final String name;
   final String address;
 
-  const WifiDirectDevice({
-    required this.name,
-    required this.address,
-  });
+  const WifiDirectDevice({required this.name, required this.address});
 }
 
-/// State of WiFi Direct sharing
-class WifiSharingState {
+class WifiServiceState {
   final String status;
   final String? address;
   final String? message;
@@ -228,7 +267,7 @@ class WifiSharingState {
   final Duration? transferTime;
   final SharedHealthPackage? receivedPackage;
 
-  const WifiSharingState._({
+  const WifiServiceState._({
     required this.status,
     this.address,
     this.message,
@@ -238,54 +277,51 @@ class WifiSharingState {
     this.receivedPackage,
   });
 
-  factory WifiSharingState.ready() => const WifiSharingState._(
-        status: 'ready',
-        message: 'Ready to share via WiFi',
-      );
+  factory WifiServiceState.ready() => const WifiServiceState._(
+    status: 'ready',
+    message: 'Ready to share via WiFi',
+  );
 
-  factory WifiSharingState.discovering() => const WifiSharingState._(
-        status: 'discovering',
-        message: 'Searching for nearby devices...',
-      );
+  factory WifiServiceState.discovering() => const WifiServiceState._(
+    status: 'discovering',
+    message: 'Searching for nearby devices...',
+  );
 
-  factory WifiSharingState.hosting(String address) => WifiSharingState._(
-        status: 'hosting',
-        address: address,
-        message: 'Waiting for connection...',
-      );
+  factory WifiServiceState.hosting(String address) => WifiServiceState._(
+    status: 'hosting',
+    address: address,
+    message: 'Waiting for connection...',
+  );
 
-  factory WifiSharingState.connecting(String address) => WifiSharingState._(
-        status: 'connecting',
-        address: address,
-        message: 'Connecting to $address...',
-      );
+  factory WifiServiceState.connecting(String address) => WifiServiceState._(
+    status: 'connecting',
+    address: address,
+    message: 'Connecting to $address...',
+  );
 
-  factory WifiSharingState.transferring(String message) => WifiSharingState._(
-        status: 'transferring',
-        message: message,
-      );
+  factory WifiServiceState.transferring(String message) =>
+      WifiServiceState._(status: 'transferring', message: message);
 
-  factory WifiSharingState.receiving() => const WifiSharingState._(
-        status: 'receiving',
-        message: 'Receiving data...',
-      );
+  factory WifiServiceState.receiving() => const WifiServiceState._(
+    status: 'receiving',
+    message: 'Receiving data...',
+  );
 
-  factory WifiSharingState.received(SharedHealthPackage package) => WifiSharingState._(
+  factory WifiServiceState.received(SharedHealthPackage package) =>
+      WifiServiceState._(
         status: 'received',
         message: 'Data received from ${package.senderNodeId}',
         receivedPackage: package,
       );
 
-  factory WifiSharingState.completed(int bytes, Duration time) => WifiSharingState._(
+  factory WifiServiceState.completed(int bytes, Duration time) =>
+      WifiServiceState._(
         status: 'completed',
         message: 'Transfer complete',
         bytesTransferred: bytes,
         transferTime: time,
       );
 
-  factory WifiSharingState.error(String message) => WifiSharingState._(
-        status: 'error',
-        message: message,
-        isError: true,
-      );
+  factory WifiServiceState.error(String message) =>
+      WifiServiceState._(status: 'error', message: message, isError: true);
 }
